+++ packages/core/e2e/transactions.test.ts
@@ -1,6 +1,6 @@
 import { Secp256k1HdWallet } from "@cosmjs/amino";
 import { FaucetClient } from "@cosmjs/faucet-client";
-import { coin, DirectSecp256k1HdWallet } from "@cosmjs/proto-signing";
+import { DirectSecp256k1HdWallet } from "@cosmjs/proto-signing";
 import {
   GasPrice,
   isDeliverTxFailure,
@@ -63,7 +63,8 @@ describe("transaction execution", () => {
       signerAddress,
       getCosmosSigner,
       getGasPrice,
-      message,
+      messages: [message],
+      chainID: "gaia-1",
     });
 
     expect(isDeliverTxSuccess(tx)).toBe(true);
@@ -109,7 +110,8 @@ describe("transaction execution", () => {
       signerAddress,
       getCosmosSigner,
       getGasPrice,
-      message,
+      messages: [message],
+      chainID: "gaia-1",
     });
 
     expect(isDeliverTxSuccess(tx)).toBe(true);
@@ -159,11 +161,11 @@ describe("transaction execution", () => {
 
     const tx = await client.executeCosmosMessage({
       signerAddress,
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore
+      // @ts-expect-error : skipping type check for testing purposes
       getCosmosSigner,
       getGasPrice,
-      message,
+      messages: [message],
+      chainID: "injective-1",
     });
 
     expect(isDeliverTxSuccess(tx)).toBe(true);
@@ -214,11 +216,11 @@ describe("transaction execution", () => {
 
     const tx = await client.executeCosmosMessage({
       signerAddress,
-      message,
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore
+      messages: [message],
+      // @ts-expect-error : skipping type check for testing purposes
       getCosmosSigner,
       getGasPrice,
+      chainID: "evmos_9000-1",
     });
 
     expect(isDeliverTxSuccess(tx)).toBe(true);
@@ -265,7 +267,8 @@ describe("transaction execution", () => {
       signerAddress,
       getCosmosSigner,
       getGasPrice,
-      message,
+      messages: [message],
+      chainID: "osmosis-1",
     });
 
     // CheckTx must pass but the execution will fail in DeliverTx due to invalid contract address
@@ -317,7 +320,8 @@ describe("transaction execution", () => {
       signerAddress,
       getCosmosSigner,
       getGasPrice,
-      message,
+      messages: [message],
+      chainID: "osmosis-1",
     });
 
     // CheckTx must pass but the execution will fail in DeliverTx due to invalid contract address
@@ -385,7 +389,7 @@ describe("transaction execution", () => {
                 destDenom:
                   "ibc/14F9BC3E44B8A9C1BE1FB08980FAB87034C9905EF17CF2F5008FC085218811CC",
                 chainID: "osmosis-1",
-                rapid_relay: true,
+                rapidRelay: true,
               },
             },
           ],
@@ -395,6 +399,7 @@ describe("transaction execution", () => {
           txsRequired: 1,
           usdAmountIn: "1.68",
           usdAmountOut: "1.68",
+          estimatedFees: [],
         },
         userAddresses: {
           "osmosis-1": signerAddress,
@@ -411,5 +416,5 @@ describe("transaction execution", () => {
       console.log(error);
       expect(error).toBeTruthy();
     }
-  });
+  }, 120000);
 });